Move init_logging functionality to the init_runtime module
